Refine fix I53e91db7 to apply only to wifi network
The original fix eliminated duplicate wifi connectivity changes stemming from
location provder scan's for APs.  These would generate two DISCONNECTED broadcasts every
two minutes and many apps mis-interpreted them.

The fix was to ignore notifications where the major state was the same as the previous one
for each network.  Unfortunately the state of per-apn notifications on cellular is hacky
and so the wifi fix was breaking mms (mms when you're on cellular with a common default+mms apn does
not generate a disconnect notification (apn still connected) so subsequent connect notifications
get dropped as duplicates).

This change refines the previous change so that it only applies to wifi networks.

bug:2392061
Change-Id: I05d8a46a4b55f8d28df8af12e05284e5e68bfc02
drno: ryanpc