Delete CM#testAccessNetworkPreference Test
Bug 3106981

ConnectivityManager#setNetworkPreference used to be synchronous,
but now it merely sends a handler message. As a result, the
method won't throw any SecurityExceptions, because it will throw
it in the handler instead when trying to write to System.Secure
settings. The test tries to write invalid settings or the same
settings value to test that it won't throw a SecurityException.
Since the method won't throw the exceptions anymore and its
not clear how long to wait to confirm that the settings weren't
going to be changed, just delete these tests.

Change-Id: I9e754bb0e026656511512b17108c42c37d84b069