Fix testVpn timing issue for com.android.cts.net.HostsideNetworkTests class
It is a issue caused by bad test case.
Please ensure the previous vpn connection is disconnected before running next
vpn test case, this will affect the test result.And the CTS test case also
tries to connect socket before the vpn rule is set in netd.
It is the same issue as http://b/18436087 .
The delay 300ms is too short, and recommend extend to 1000 or 3000 ms.
Since L MR1 is known as supporting some low memory device, we suggest fixing
the new CTS test case for some low-end devices.

Change-Id: I6051c95c663e867fc66c580ce8bd8d25fce0b8fc
Signed-off-by: Junjie Hu <junjie.hu@mediatek.com>