Remove non-CTS packets assertion code
Symptom:
testTrafficStatsForLocalhost fails when applications or services
except CTS are using network.

Root cause:
testTrafficStatsForLocalhost can't calculate the number of
non-localhost packets for CTS.
It includes the all of packets for applications/services except CTS
and non-localhost packets for CTS.

Solution:
Remove this assertion and only logging the number of packets.

Change-Id: I49243d59f359f3a543c6bdb46f6a2645cde8f292